set default logger level to error, and give controll method
The output of default log are noisy, set to error default.

Give an method to controll at program startup time. To set level to info,
set enviroment like this:
    export RPCX_LOG_LEVEL=4
To turn everything on,
    export RPCX_LOG_LEVEL=99
the value of levels are:
	- lvPicni = 0
	- lvFatal = 1
	- lvError = 2
	- lvWarn = 3
	- lvInfo = 4
	- lvDebug = 5

This trick works on my company, you may interested. If not, just ignore
and drop this pr.